A bunch of changes but it works
