+++ version_b/main.rb
@@ -2,7 +2,6 @@ require_relative 'mission_reporter'
 require_relative 'mission_control'
 require_relative 'mission_plan'
 require_relative 'mission'
-require 'pry'
 
 mission_collections = MissionControl.new(mission_instance: Mission.new)
 puts 'Welcome to yellow-steel-freetown!'
