+++ version_b/mission_reporter.rb
@@ -1,23 +1,26 @@
 class MissionReporter
+  attr_reader :mission
+
   def initialize(mission)
     @mission = mission
   end
+
   def print_status
     puts 'Mission status:'
     puts "  Current fuel burn rate: #{Mission::BURN_RATE} liters/min"
-    puts "  Current speed: #{(@mission.current_speed * Mission::SECONDS_PER_HOURS).round(2)} km/h"
-    puts "  Elapsed time: #{seconds_to_hms(@mission.elapsed_time)}"
-    puts "  Distance traveled: #{@mission.distance_traveled.round(2)} km"
-    puts "  Time to destination: #{@mission.time_to_destination.round(2)} seconds"
+    puts "  Current speed: #{(mission.current_speed * Mission::SECONDS_PER_HOURS).round(2)} km/h"
+    puts "  Elapsed time: #{seconds_to_hms(mission.elapsed_time)}"
+    puts "  Distance traveled: #{mission.distance_traveled.round(2)} km"
+    puts "  Time to destination: #{mission.time_to_destination.round(2)} seconds"
   end
 
   def print_summary
     puts "Mission summary:"
-    puts "  Total distance traveled: #{distance_traveled.round(2)} km"
-    puts "  Number of aborts and retries: #{@aborts}/#{@retries}"
-    puts "  Number of explosions: #{@explosions}"
-    puts "  Total fuel burned: #{total_fuel_burned.round(0)} liters"
-    puts "  Flight time: #{seconds_to_hms(total_time)}"
+    puts "  Total distance traveled: #{mission.distance_traveled.round(2)} km"
+    puts "  Number of aborts and retries: #{mission.aborts}/#{mission.retries}"
+    puts "  Number of explosions: #{mission.explosions}"
+    puts "  Total fuel burned: #{mission.total_fuel_burned.round(0)} liters"
+    puts "  Flight time: #{seconds_to_hms(mission.total_time)}"
   end
 
   private
